Move dtype utilities to utils/dtypes.py
