+++ python/cudf/cudf/utils/utils.py
@@ -1,5 +1,3 @@
-import numbers
-from collections import namedtuple
 from math import ceil, isinf, isnan
 
 import numpy as np
@@ -8,8 +6,6 @@ import pyarrow as pa
 from numba import njit
 
 from librmm_cffi import librmm as rmm
-
-from cudf.utils.dtypes import is_categorical_dtype
 
 mask_dtype = np.dtype(np.int8)
 mask_bitsize = mask_dtype.itemsize * 8
@@ -20,19 +16,6 @@ def calc_chunk_size(size, chunksize):
     return mask_byte_padding * ceil(
         ((size + chunksize - 1) // chunksize) / mask_byte_padding
     )
-
-
-_TypeMinMax = namedtuple("_TypeMinMax", "min,max")
-
-
-def get_numeric_type_info(dtype):
-    if dtype.kind in "iu":
-        info = np.iinfo(dtype)
-        return _TypeMinMax(info.min, info.max)
-    elif dtype.kind in "f":
-        return _TypeMinMax(dtype.type("-inf"), dtype.type("+inf"))
-    else:
-        raise TypeError(dtype)
 
 
 @njit
@@ -74,6 +57,7 @@ def require_writeable_array(arr):
 
 def scalar_broadcast_to(scalar, shape, dtype):
     from cudf.utils.cudautils import fill_value
+    from cudf.utils.dtypes import to_cudf_compatible_scalar
 
     scalar = to_cudf_compatible_scalar(scalar, dtype=dtype)
 
@@ -139,102 +123,6 @@ def buffers_from_pyarrow(pa_arr, dtype=None):
     else:
         padata = Buffer(np.empty(0, dtype=new_dtype))
     return (pamask, padata)
-
-
-def cudf_dtype_from_pydata_dtype(dtype):
-    """ Given a numpy or pandas dtype, converts it into the equivalent cuDF
-        Python dtype.
-    """
-    try:
-        # pd 0.24.X
-        from pandas.core.dtypes.common import infer_dtype_from_object
-    except ImportError:
-        # pd 0.23.X
-        from pandas.core.dtypes.common import (
-            _get_dtype_from_object as infer_dtype_from_object,
-        )
-
-    if is_categorical_dtype(dtype):
-        pass
-    elif np.issubdtype(dtype, np.datetime64):
-        dtype = np.datetime64
-
-    return infer_dtype_from_object(dtype)
-
-
-def is_scalar(val):
-    return (
-        val is None
-        or isinstance(val, str)
-        or isinstance(val, numbers.Number)
-        or np.isscalar(val)
-        or isinstance(val, pd.Timestamp)
-        or isinstance(val, pd.Categorical)
-    )
-
-
-def to_cudf_compatible_scalar(val, dtype=None):
-    """
-    Converts the value `val` to a numpy/Pandas scalar,
-    optionally casting to `dtype`.
-
-    If `val` is None, returns None.
-    """
-    if val is None:
-        return val
-
-    if not is_scalar(val):
-        raise ValueError(
-            f"Cannot convert value of type {type(val).__name__} "
-            " to cudf scalar"
-        )
-
-    val = pd.api.types.pandas_dtype(type(val)).type(val)
-
-    if dtype is not None:
-        val = val.astype(dtype)
-
-    return val
-
-
-def is_list_like(obj):
-    """
-    This function checks if the given `obj`
-    is a list-like (list, tuple, Series...)
-    type or not.
-
-    Parameters
-    ----------
-    obj : object of any type which needs to be validated.
-
-    Returns
-    -------
-    Boolean: True or False depending on whether the
-    input `obj` is like-like or not.
-    """
-    from collections.abc import Sequence
-
-    if isinstance(obj, (Sequence,)) and not isinstance(obj, (str, bytes)):
-        return True
-    else:
-        return False
-
-
-def min_scalar_type(a, min_size=8):
-    return min_signed_type(a, min_size=min_size)
-
-
-def min_signed_type(x, min_size=8):
-    """
-    Return the smallest *signed* integer dtype
-    that can represent the integer ``x``
-    """
-    for int_dtype in np.sctypes["int"]:
-        if (np.dtype(int_dtype).itemsize * 8) >= min_size:
-            if np.iinfo(int_dtype).min <= x <= np.iinfo(int_dtype).max:
-                return int_dtype
-    # resort to using `int64` and let numpy raise appropriate exception:
-    return np.int64(x).dtype
 
 
 def get_result_name(left, right):
